pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinelpi"
version = "1.0.0"
description = "Lightweight defensive network anomaly monitor for Raspberry Pi"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "SentinelPi Project"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Networking :: Monitoring",
]
dependencies = [
    "scapy>=2.5.0,<3.0",
    "psutil>=5.9.0,<6.0",
    "flask>=3.0.0,<4.0",
    "pyyaml>=6.0,<7.0",
    "requests>=2.31.0,<3.0",
    "python-dateutil>=2.8.0,<3.0",
]

[project.optional-dependencies]
geo = ["maxminddb>=2.4.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
sentinelpi = "sentinelpi.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 120
